feat(properties): Enhance property and user management
- Update BienSummaryResponse DTO with additional fields
- Add BienUpdateRequest DTO for property updates
- Update PhotoBien entity with new properties
- Enhance BienService with improved business logic
- Update UserDto and UserService with new features

+++ backend/src/main/java/com/immoteam/immoteamdev/dto/BienUpdateRequest.java
@@ -12,6 +12,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -32,11 +34,11 @@ public class BienUpdateRequest {
 
     @NotNull(message = "Le prix est obligatoire")
     @DecimalMin(value = "0.0", inclusive = false, message = "Le prix doit être supérieur à 0")
-    private Double prix;
+    private BigDecimal prix;
 
     @NotNull(message = "La surface est obligatoire")
     @DecimalMin(value = "0.0", inclusive = false, message = "La surface doit être supérieure à 0")
-    private Double surface;
+    private BigDecimal surface;
 
     @Min(value = 0, message = "Le nombre de chambres ne peut pas être négatif")
     private Integer nbChambres;
